add: create a VM and move funcs out
